pyproject.toml
[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
line_length = 79
profile = "black"
multi_line_output = 3

[tool.poetry]
name = "smartshala-backend"
version = "0.1.0"
description = ""
authors = ["Ananyobrata Pal <ananyo141@gmail.com>"]
license = "GPLv3"
# readme = "README.md"
packages = [{include = "smartshala"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.14.0"
python-dotenv = "^1.0.0"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
drf-spectacular = "^0.26.0"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pre-commit = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
